edit starting, heading quick fix input side, some other documentation -> entry_collection.js, next-> addnewaction(controller)
