+++ app/Resources/views/layout.html.twig
@@ -8,6 +8,7 @@
 
   {% block stylesheets %}
     <link href="//maxcdn.bootstrapcdn.com/bootstrap/3.3.4/css/bootstrap.min.css" rel="stylesheet">
+    <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/font-awesome/4.5.0/css/font-awesome.min.css">
     {% stylesheets '../app/Resources/public/css/*' %}
       <link rel="stylesheet" type="text/css" charset="UTF-8" media="all" href="{{ asset_url }}"/>
     {% endstylesheets %}
