+++ src/SE/InputBundle/Resources/views/Entry/input_form.html.twig
@@ -48,10 +48,7 @@
       </a>
       <div class="hide toggling txtarea-sm">
         {{ form_widget(entry.comments, {'attr': {'class': 'form-control'}}) }}
-      </div>
-      <div class="hide toggling input-manual">
-        <p class="pull-left">Manual TO :</p> {{ form_widget(entry.totalTo, {'attr': {'class': 'form-control'}}) }}
-      </div>   
+      </div>  
     </td>
     {{ form_widget(entry) }}
   </tr>
