+++ src/SE/InputBundle/Resources/views/Import/set_import.html.twig
@@ -4,19 +4,23 @@
   SAP Import Settings- {{ parent() }}
 {% endblock %}
 
-{% block seinput_body %}
-
+{% block seinput_header %}
+  
   <header class="h-sap">
   	<div class="line"></div>
   	<h2><i class="glyphicon glyphicon-cloud"> </i> SAP import</h2>
   </header>
 
-	<div class="well col-md-6">
+{% endblock %}
+
+{% block seinput_body %}
+
+	<div class="well col-md-10">
 		<form class="form-inline" method="post" action="http://localhost/SE/Import/Import.php"> 
    			<div class="form-group">
 			    <label for="dateinput">Date of confirmation : </label>
 			    <div class="input-group">
-			    	<input type="date" class="form-control" id="dateinput" value="{{ 'now'|date('d/m/Y') }}" name="date">
+			    	<input type="date" class="form-control" id="dateinput" value="{{ 'now'|date_modify("-1 day")|date('d/m/Y') }}" name="date">
 				    <span class="input-group-addon">
 	          			<span id="overtime" class="glyphicon glyphicon-time"></span>
 			        </span>
@@ -77,14 +81,13 @@
 
 {% block javascripts %}
   {{ parent() }}
-  <script type="text/javascript" charset="utf8" src="//cdn.datatables.net/1.10.7/js/jquery.dataTables.js"></script>
   <script type="text/javascript">
   	$('#imports').DataTable( {
 		"paging": true,
 	  	"dom": 'lrtip',
-	  	"info": false
+	  	"info": false,
+	  	"order": [[ 0, "desc" ]]
 	});
-
   </script>
   
 {% endblock %}
