+++ src/SE/InputBundle/Resources/views/layout.html.twig
@@ -8,17 +8,20 @@
   {{ render(controller("SEInputBundle:Entry:menu")) }}
 {% endblock %}
 
-{% block stylesheets %}
-  {{ parent() }}
-  {% stylesheets 'bundles/seinput/css/input_style.css' %}
-    <link rel="stylesheet" type="text/css" charset="UTF-8" media="all" href="{{ asset_url }}"/>
-  {% endstylesheets %}
-{% endblock %}
-
 {% block body %}
 
-  {# On définit un nouveau bloc, que les vues du bundle pourront remplir #}
+  <div class="header-block">
+	{% block seinput_header %}
+    {% endblock %}
+  </div>
+
   {% block seinput_body %}
   {% endblock %}
 
+{% endblock %}
+
+{% block javascripts %}
+  	{{ parent() }}
+	<script type="text/javascript" charset="utf8" src="https://cdn.datatables.net/s/bs/jszip-2.5.0,pdfmake-0.1.18,dt-1.10.10,b-1.1.0,b-html5-1.1.0,b-print-1.1.0/datatables.min.js"></script>
+  	<script src="http://code.highcharts.com/highcharts.js"></script>
 {% endblock %}
